+++ order-service/src/main/java/com/palja/order_service/application/dto/response/DeliveryRes.java
@@ -1,4 +1,4 @@
-package com.palja.order_service.application.dto;
+package com.palja.order_service.application.dto.response;
 
 import com.palja.order_service.domain.entity.OrderDelivery;
 import lombok.AccessLevel;
