+++ order-service/src/main/java/com/palja/order_service/application/service/OrderManagerService.java
@@ -1,7 +1,7 @@
 package com.palja.order_service.application.service;
 
 import com.palja.order_service.application.command.ChangeOrderStatusCommand;
-import com.palja.order_service.application.dto.OrderStatusChangeRes;
+import com.palja.order_service.application.dto.response.OrderStatusChangeRes;
 
 public interface OrderManagerService {
     OrderStatusChangeRes changeOrderStatus(ChangeOrderStatusCommand command);
