+++ order-service/src/main/java/com/palja/order_service/application/service/OrderService.java
@@ -3,9 +3,9 @@ package com.palja.order_service.application.service;
 import com.palja.common.vo.UserRole;
 import com.palja.order_service.application.command.CancelOrderCommand;
 import com.palja.order_service.application.command.CreateOrderCommand;
-import com.palja.order_service.application.dto.OrderCancelRes;
-import com.palja.order_service.application.dto.OrderCreateRes;
-import com.palja.order_service.application.dto.OrderDetailRes;
+import com.palja.order_service.application.dto.response.OrderCancelRes;
+import com.palja.order_service.application.dto.response.OrderCreateRes;
+import com.palja.order_service.application.dto.response.OrderDetailRes;
 import com.palja.order_service.domain.entity.Order;
 
 import java.util.UUID;
