+++ order-service/src/main/java/com/palja/order_service/application/service/validator/OrderValidator.java
@@ -4,7 +4,7 @@ import com.palja.common.exception.BusinessException;
 import com.palja.common.vo.UserRole;
 import com.palja.order_service.application.command.CreateOrderCommand;
 import com.palja.order_service.application.command.DeliveryCommand;
-import com.palja.order_service.application.dto.*;
+import com.palja.order_service.application.dto.response.*;
 import com.palja.order_service.application.exception.OrderErrorCode;
 import com.palja.order_service.application.service.UserService;
 import com.palja.order_service.domain.entity.Order;
