+++ order-service/src/main/java/com/palja/order_service/presentation/controller/OrderController.java
@@ -4,9 +4,9 @@ import com.palja.common.annotation.RequiredRole;
 import com.palja.common.auditor.CurrentUser;
 import com.palja.common.response.ApiResponse;
 import com.palja.common.vo.UserRole;
-import com.palja.order_service.application.dto.OrderCancelRes;
-import com.palja.order_service.application.dto.OrderCreateRes;
-import com.palja.order_service.application.dto.OrderDetailRes;
+import com.palja.order_service.application.dto.response.OrderCancelRes;
+import com.palja.order_service.application.dto.response.OrderCreateRes;
+import com.palja.order_service.application.dto.response.OrderDetailRes;
 import com.palja.order_service.application.service.OrderService;
 import com.palja.order_service.presentation.dto.request.CancelOrderReq;
 import com.palja.order_service.presentation.dto.request.CreateOrderReq;
